Revert "fabtests: rename files and directories to better reflect intent"
This reverts commit 4c579b65055ec3145f47c071b68b1c21446b6ed5.

Revering this commit because naming a directory "core" creates
problems with Autoconf built-in macros that execute "rm -f core ..."
(because they're trying to clean up stale corefiles).  A variety of
minor and major errors occur in configure because "rm -f core" fails
to remove the directory, and therefore errors out.  A noticable error
that occurs is that configure leaves "confdefs.h" in the top-level
directory, which both causes "make distcheck" to fail and leaves a
bunch of read-only distcheck files in the build tree.

NOTE: Originally, 4c579b6505 renamed the "fabtests/common" directory
to "fabtests/core".  Subsequently, commit d56870751 modified files in
the new "fabtests/core" directory.

It looks like "git revert" was smart enough to revert 4c579b6505 but
still retain the changes from d56870751.  I.e., "fabtests/common"
appears to have the changes from d56870751.
