+++ fabtests/benchmarks/dgram_pingpong.c
@@ -38,7 +38,7 @@
 #include <rdma/fi_errno.h>
 #include <rdma/fi_endpoint.h>
 
-#include "common.h"
+#include "shared.h"
 #include "benchmark_shared.h"
 
 static int run(void)
