+++ fabtests/benchmarks/rdm_cntr_pingpong.c
@@ -33,7 +33,7 @@
 
 #include <rdma/fi_errno.h>
 
-#include <common.h>
+#include <shared.h>
 #include "benchmark_shared.h"
 
 static int run(void)
