+++ fabtests/common/windows/getopt.c
