+++ fabtests/common/windows/osd.c
