+++ fabtests/functional/cq_data.c
@@ -35,7 +35,7 @@
 #include <rdma/fi_endpoint.h>
 #include <rdma/fi_cm.h>
 
-#include <common.h>
+#include <shared.h>
 
 static int run_test()
 {
