+++ fabtests/functional/mcast.c
@@ -32,7 +32,7 @@
 #include <getopt.h>
 #include <rdma/fi_cm.h>
 
-#include <common.h>
+#include <shared.h>
 
 
 static int listener;
