+++ fabtests/functional/multi_mr.c
@@ -42,7 +42,7 @@
 #include <rdma/fi_rma.h>
 #include <rdma/fi_cm.h>
 
-#include "common.h"
+#include "shared.h"
 
 struct test_mr {
 	uint8_t *buf;
