+++ fabtests/functional/rdm_deferred_wq.c
@@ -36,7 +36,7 @@
 #include <rdma/fi_errno.h>
 #include <rdma/fi_trigger.h>
 
-#include <common.h>
+#include <shared.h>
 
 struct fi_context2 trig_ctx1, trig_ctx2;
 static char *welcome_text = "Hello from Client!";
