+++ fabtests/functional/rdm_multi_domain.c
@@ -43,7 +43,7 @@
 #include <rdma/fi_rma.h>
 #include <rdma/fi_cm.h>
 
-#include "common.h"
+#include "shared.h"
 
 struct test_domain {
 	uint8_t *buf;
