+++ fabtests/functional/recv_cancel.c
@@ -32,7 +32,7 @@
 
 #include <stdio.h>
 #include <unistd.h>
-#include "common.h"
+#include "shared.h"
 
 #define CANCEL_TAG 0xC
 #define STANDARD_TAG 0xA
