+++ fabtests/functional/shared_ctx.c
@@ -37,7 +37,7 @@
 #include <rdma/fi_endpoint.h>
 #include <rdma/fi_cm.h>
 
-#include <common.h>
+#include <shared.h>
 
 enum {
 	FT_UNSPEC,
