+++ fabtests/functional/unexpected_msg.c
@@ -46,7 +46,7 @@
 #include <rdma/fi_rma.h>
 #include <rdma/fi_cm.h>
 
-#include "common.h"
+#include "shared.h"
 
 
 static size_t concurrent_msgs = 5;
