+++ fabtests/include/shared.h
