+++ fabtests/unit/av_test.c
@@ -41,7 +41,7 @@
 
 #include <rdma/fi_errno.h>
 
-#include "common.h"
+#include "shared.h"
 #include "unit_common.h"
 
 #define MAX_ADDR 256
